Move `iota_iterator_into its own file`
